DOCS: Fix counters in Firefox (#177)
* update to counter for Firefox

Firefox doesn't seem to support `counter-reset` if the parent also has a `counter-reset`.

Instead, the parent must use `counter-set`, and then FF will honor the reset on the children. Still works in other browsers.

This had an impact on the colors. When you hover on them in FF, because the reset wasn't carrying over, the numbering would go from 1 to 10, rather than 0 to 9.

* different approach for better browser support

* make the fix work

* Fix layout issues caused by the pseudo-element