durations.css – correct leap-year duration; introduce olympiad (#475)
Hi there,

just noted an error in the durations. A leap year does not refer to a time span of four years, but to the length of the year where the leap day occurs, i.e. 366 days.

On the other hand, conveniently, there *is* a duration that describes a four-year timespan, an olympiad, which I introduced as a new property.